requirements.txt
accept==0.1.0
alchemyjsonschema==0.2.2
antiorm==1.2.0
docker-py==1.7.2
jsl==0.2.2
jsonpatch==1.12
jsonpointer==1.10
jsonschema==2.3.0
loremipsum==1.0.5
Momoko==2.2.3
pytest==2.9.1
pytest-pep8==1.0.6
python_jsonschema_objects==0.0.15
python-status==1.0.1
rfc3987==1.3.5
SQLAlchemy==1.0.12
tornado>=4.2
WebTest==2.0.20
